pipeline bits refactored inside Pipeline/*.js
